refactor: remove app history reader
